Make FormData#getComputedLength implementation syncronous, but keep the function itself async.
Since after the latest refactoring getLength helper became synchronouse, the `FormData#getComputedLength()` method can now be synchronous as well, so
the public API for that method (only its returning type) must be changed in a future, but I'm not sure whether this counts as major semver change or not.
I'll keep it as async-ish function for not, until I decide how to handle this change properly.